version 4.3
1. The brand-new pipeline V2, which decompose the structure (depth) and texture (rgb), and achieve backward warp. Which naturally support flow supervision in occluded region.
2. Scripts to prepare MannequinChallenge. and change pytube to youtube-dl

TODO:
try different settings of the pipeline v2 and check problem in scene flow?

+++ script/prepare_RealEstate10K.py
@@ -12,12 +12,13 @@ from glob import glob
 sys.path.append("..")
 from dataset.RealEstate10K import RealEstate10K_Img, RealEstate10K_root
 
-np.random.seed(0)
+# np.random.seed(0)
 process_train = True
 
 train_str = "train" if process_train else "test"
 success_file = os.path.join(RealEstate10K_root, f"{train_str}_valid.txt")
 black_list_name = os.path.join(RealEstate10K_root, "black_list.txt")
+prepare_black_list_name = os.path.join(RealEstate10K_root, "prepare_black_list.txt")
 file_bases = glob(os.path.join(RealEstate10K_root, train_str, "*.txt"))
 file_bases = [os.path.basename(fb).split('.')[0] for fb in file_bases]
 
@@ -26,8 +27,12 @@ with open(success_file, 'r') as file:
     lines = file.readlines()
 with open(black_list_name, 'r') as file:
     black_lines = file.readlines()
+with open(prepare_black_list_name, 'r') as file:
+    prepare_black_list = file.readlines()
 lines = {line.strip('\n') for line in lines}
 black_lines = {os.path.basename(line.strip('\n')) for line in black_lines}
+prepare_black_list = {os.path.basename(line.strip('\n')) for line in prepare_black_list}
+black_lines.update(prepare_black_list)
 
 trainset = RealEstate10K_Img(process_train)
 
@@ -41,14 +46,17 @@ for train_id in train_seq:
     print(f"{len(lines)}/{len(file_bases)} ({train_failed_num} failed): {(time.time() - starttime):.1f}s"
           f" at {datetime.datetime.now()}")
     starttime = time.time()
-    try:
-        ret = trainset.getitem_bybase(file_bases[train_id])
-    except Exception as e:
-        ret = None
-        print(e)
+
     name = file_bases[train_id]
     if os.path.basename(name) in black_lines:
+        print(f"{os.path.basename(name)} in black_list")
         ret = None
+    else:
+        try:
+            ret = trainset.getitem_bybase(name)
+        except Exception as e:
+            ret = None
+            print(e)
 
     if ret is not None:
         if name in lines:
@@ -70,6 +78,7 @@ for train_id in train_seq:
             file_base = os.path.basename(name).split('.')[0]
             dir_base = os.path.join(trainset.tmp_root, file_base)
             print(f"{os.path.basename(name)} failed, deleting {dir_base}")
-            shutil.rmtree(dir_base)
+            if os.path.exists(dir_base):
+                shutil.rmtree(dir_base)
     print("-----------------------------------------")
